# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LUFFY XO.SHOP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  password: string;
  username: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  userId: string;
  email: string;
  username: string;
  balance: number;
  isAdmin: boolean;
  /** @nullable */
  lastLoginAt: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateBalanceBody {
  balance: number;
}

export interface Transaction {
  id: number;
  userId: number;
  type: string;
  amount: number;
  status: string;
  /** @nullable */
  reference: string | null;
  /** @nullable */
  note: string | null;
  createdAt: string;
  user: User;
}

export interface TopupBody {
  amount: number;
  reference: string;
}

export interface TxtFile {
  id: number;
  name: string;
  totalLines: number;
  createdAt: string;
}

export interface UploadTxtFileBody {
  name: string;
  content: string;
}

export interface GenerateTxtBody {
  fileId: number;
  lineCount: number;
}

export interface GenerateTxtResponse {
  lines: string[];
  coinsSpent: number;
  remainingBalance: number;
}

export interface GenerationHistory {
  id: number;
  fileId: number;
  fileName: string;
  lineCount: number;
  coinsSpent: number;
  createdAt: string;
}

export interface GenerateCodmBody {
  count: number;
}

export interface CodmAccount {
  id: number;
  account: string;
  status: string;
  coinsSpent: number;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalCoinsDistributed: number;
  totalGenerations: number;
  pendingTopups: number;
  totalRevenue: number;
  totalCodmGenerated: number;
}

export interface LoginEvent {
  id: number;
  userId: number;
  username: string;
  email: string;
  loginAt: string;
}